Remove some unnecessary code
