+++ package.json
@@ -13,8 +13,8 @@
     "lint": "eslint scripts/ packages/ docs/ --ext js,md",
     "format": "prettier \"*.{js,md,json}\" \"{docs,packages,scripts}/**/*.{js,md,json}\" --write",
     "format:ci": "yarn format -- --list-different",
-    "test": "cross-env NODE_ENV=test VERSION=\"1\" karma start --single-run",
-    "test:watch": "cross-env NODE_ENV=test VERSION=\"1\" karma start",
+    "test": "karma start --single-run",
+    "test:watch": "karma start",
     "check-snapshots": "lerna run check-snapshot",
     "posttest": "[ -z \"$TRAVIS\" ] || codecov",
     "codecov": "codecov",
@@ -103,7 +103,7 @@
     "rollup-plugin-size-snapshot": "^0.6.1",
     "rollup-plugin-uglify": "^4.0.0",
     "shelljs": "^0.8.2",
-    "sinon": "4.5.0",
+    "sinon": "^7.2.2",
     "typescript": "^3.1.3",
     "webpack": "^2.3.3",
     "zen-observable": "^0.6.0"
