--- packages/jss/tests/.babelrc
@@ -1,10 +1,0 @@
-{
-  "extends": "../../../.babelrc",
-  "env": {
-    "test": {
-      "plugins": [
-        "transform-runtime"
-      ]
-    }
-  }
-}
