--- packages/jss/tests/index.js
@@ -1,17 +1,0 @@
-import './utils'
-import './integration/rules'
-import './integration/sheet'
-import './integration/sheetsRegistry'
-import './integration/plugins'
-import './integration/houdini'
-import './functional/rules'
-import './functional/sheet'
-import './functional/priority'
-import './functional/houdini'
-import './unit/getDynamicStyles'
-import './unit/cloneStyle'
-import './unit/SheetsManager'
-import './unit/createGenerateId'
-import './unit/escape'
-import './unit/Jss'
-import './unit/plugins'
