--- tests.webpack.js
@@ -1,2 +1,0 @@
-const context = require.context('./tests', true, /\.js$/)
-context.keys().forEach(context)
